requirements.txt
Pillow
deep-translator
pywin32
easyocr
google-generativeai
python-dotenv
